pyproject.toml
[tool.poetry]
name = "contacts-book"
version = "0.1.0"
description = ""
authors = ["Oleksii Hladkov <agladkoff17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
sqlalchemy = "^2.0.23"
psycopg2 = "^2.9.9"
alembic = "^1.13.0"
pydantic = {extras = ["email"], version = "^2.5.2"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
fastapi-mail = "^1.4.1"
fastapi-limiter = "^0.1.5"
cloudinary = "^1.37.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"
pytest = "^7.4.3"
httpx = "^0.25.2"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
